add mongoose validation and error handling

+++ src/controllers/movieController.js
@@ -1,5 +1,4 @@
 import { Router } from "express";
-
 import movieService from "../services/movieService.js";
 import castService from "../services/castService.js";
 import { isAuth } from "../middlewares/authMiddleware.js"
@@ -15,8 +14,12 @@ router.get('/create', isAuth, (req, res) => {
 router.post('/create', isAuth, async (req, res) => {
     const movieData = req.body;
     const ownerId = req.user?._id;
-    await movieService.create(movieData, ownerId);
-
+    try {
+       await movieService.create(movieData, ownerId); 
+    } catch (err) {
+        const errorMessage = Object.values(err.errors)[0]?.message;
+        return res.render('movies/create', { error: errorMessage, movie: movieData}); 
+    }
     res.redirect('/');
 });
 
@@ -31,9 +34,6 @@ router.get('/:movieId/details', async (req, res) => {
     const movieId = req.params.movieId;
     const movie = await movieService.getOne(movieId).lean(); 
     const isOwner = movie.owner && movie.owner.toString() === req.user?._id;
-
-    
-
     res.render('movies/details', { movie, isOwner});
 });
 
@@ -56,7 +56,15 @@ router.post('/:movieId/attach', isAuth, async (req, res) => {
 
 router.get('/:movieId/delete', isAuth, async (req,res) => {
    const movieId = req.params.movieId;
+
+   // check if is the owner if is not you can not delete the movie!
+   const movie = await movieService.getOne(movieId).lean();
+   if (movie.owner.toString() !== req.user._id) {
+    return res.redirect(`/movies/${movieId}/details`);
+   }
+
    await movieService.remove(movieId);
+
    res.redirect('/');
 });
 
@@ -73,4 +81,4 @@ router.post('/:movieId/edit', isAuth, async (req,res) => {
    res.redirect(`/movies/${movieId}/details`);
 });
 
-export default router;
+export default router;
